+++ serve/response.go
@@ -1,11 +1,11 @@
 package serve
 
 type Response struct {
-	Status uint64                 `json:"status"`
+	Status int                    `json:"status"`
 	Data   map[string]interface{} `json:"data"`
 }
 
-func NewResponse(status uint64, data map[string]interface{}) *Response {
+func NewResponse(status int, data map[string]interface{}) *Response {
 	return &Response{
 		Status: status,
 		Data:   data,
